# docs/customizations/openapi.json
{
  "swagger": "2.0",
  "info": {
    "title": "Mayan EDMS API",
    "description": "Free Open Source Electronic Document Management System",
    "license": {
      "name": "Apache 2.0"
    },
    "version": "v4"
  },
  "host": "localhost:8080",
  "schemes": [
    "http"
  ],
  "basePath": "/api/v4",
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "securityDefinitions": {
    "Basic": {
      "type": "basic"
    }
  },
  "security": [
    {
      "Basic": []
    }
  ],
  "paths": {
...
    "/cabinets/{cabinet_id}/documents/": {
      "get": {
        "operationId": "cabinets_documents_list",
        "description": "Returns a list of all the documents contained in a particular cabinet.",
        "parameters": [
          {
            "name": "_ordering",
            "in": "query",
            "description": "Which field to use when ordering the results.",
            "required": false,
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "description": "A page number within the paginated result set.",
            "required": false,
            "type": "integer"
          },
          {
            "name": "page_size",
            "in": "query",
            "description": "Number of results to return per page.",
            "required": false,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "cabinets"
        ]
      },
      "parameters": [
        {
          "name": "cabinet_id",
          "in": "path",
          "required": true,
          "type": "string"
        }
      ]
    },
...
    "/document_types/": {
      "get": {
        "operationId": "document_types_list",
        "description": "Returns a list of all the document types.",
        "parameters": [
          {
            "name": "_ordering",
            "in": "query",
            "description": "Which field to use when ordering the results.",
            "required": false,
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "description": "A page number within the paginated result set.",
            "required": false,
            "type": "integer"
          },
          {
            "name": "page_size",
            "in": "query",
            "description": "Number of results to return per page.",
            "required": false,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "document_types"
        ]
      },
      "post": {
        "operationId": "document_types_create",
        "description": "Create a new document type.",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "document_types"
        ]
      },
      "parameters": []
    },
    "/document_types/{document_type_id}/": {
      "get": {
        "operationId": "document_types_read",
        "description": "Return the details of the selected document type.",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "document_types"
        ]
      },
      "put": {
        "operationId": "document_types_update",
        "description": "Edit the properties of the selected document type.",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "document_types"
        ]
      },
      "patch": {
        "operationId": "document_types_partial_update",
        "description": "Edit the properties of the selected document type.",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "document_types"
        ]
      },
      "delete": {
        "operationId": "document_types_delete",
        "description": "Delete the selected document type.",
        "parameters": [],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "document_types"
        ]
      },
      "parameters": [
        {
          "name": "document_type_id",
          "in": "path",
          "required": true,
          "type": "string"
        }
      ]
    },
...
    "/documents/": {
      "get": {
        "operationId": "documents_list",
        "description": "Returns a list of all the documents.",
        "parameters": [
          {
            "name": "_ordering",
            "in": "query",
            "description": "Which field to use when ordering the results.",
            "required": false,
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "description": "A page number within the paginated result set.",
            "required": false,
            "type": "integer"
          },
          {
            "name": "page_size",
            "in": "query",
            "description": "Number of results to return per page.",
            "required": false,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "documents"
        ]
      },
      "post": {
        "operationId": "documents_create",
        "description": "Create a new document.",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "documents"
        ]
      },
      "parameters": []
    },
...
    "/documents/{document_id}/": {
      "get": {
        "operationId": "documents_read",
        "description": "Return the details of the selected document.",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "documents"
        ]
      },
      "put": {
        "operationId": "documents_update",
        "description": "Edit the properties of the selected document.",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "documents"
        ]
      },
      "patch": {
        "operationId": "documents_partial_update",
        "description": "Edit the properties of the selected document.",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "documents"
        ]
      },
      "delete": {
        "operationId": "documents_delete",
        "description": "Move the selected document to the thrash.",
        "parameters": [],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "documents"
        ]
      },
      "parameters": [
        {
          "name": "document_id",
          "in": "path",
          "required": true,
          "type": "string"
        }
      ]
    },
...
    "/documents/{document_id}/resolved_smart_links/{resolved_smart_link_id}/documents/": {
      "get": {
        "operationId": "documents_resolved_smart_links_documents_list",
        "description": "Returns a list of the smart link documents that apply to the document.",
        "parameters": [
          {
            "name": "_ordering",
            "in": "query",
            "description": "Which field to use when ordering the results.",
            "required": false,
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "description": "A page number within the paginated result set.",
            "required": false,
            "type": "integer"
          },
          {
            "name": "page_size",
            "in": "query",
            "description": "Number of results to return per page.",
            "required": false,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "documents"
        ]
      },
      "parameters": [
        {
          "name": "document_id",
          "in": "path",
          "required": true,
          "type": "string"
        },
        {
          "name": "resolved_smart_link_id",
          "in": "path",
          "required": true,
          "type": "string"
        }
      ]
    },
...
    "/index_instances/{index_instance_id}/nodes/{index_instance_node_id}/documents/": {
      "get": {
        "operationId": "index_instances_nodes_documents_list",
        "description": "Returns a list of all the documents contained by a particular\nindex instance node.",
        "parameters": [
          {
            "name": "_ordering",
            "in": "query",
            "description": "Which field to use when ordering the results.",
            "required": false,
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "description": "A page number within the paginated result set.",
            "required": false,
            "type": "integer"
          },
          {
            "name": "page_size",
            "in": "query",
            "description": "Number of results to return per page.",
            "required": false,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "index_instances"
        ]
      },
      "parameters": [
        {
          "name": "index_instance_id",
          "in": "path",
          "required": true,
          "type": "string"
        },
        {
          "name": "index_instance_node_id",
          "in": "path",
          "required": true,
          "type": "string"
        }
      ]
    },
...
    "/index_templates/{index_template_id}/document_types/": {
      "get": {
        "operationId": "index_templates_document_types_list",
        "description": "Returns a list of the document types associated with this index template.",
        "parameters": [
          {
            "name": "_ordering",
            "in": "query",
            "description": "Which field to use when ordering the results.",
            "required": false,
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "description": "A page number within the paginated result set.",
            "required": false,
            "type": "integer"
          },
          {
            "name": "page_size",
            "in": "query",
            "description": "Number of results to return per page.",
            "required": false,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "index_templates"
        ]
      },
      "parameters": [
        {
          "name": "index_template_id",
          "in": "path",
          "required": true,
          "type": "string"
        }
      ]
    },
...
    "/smart_links/{smart_link_id}/document_types/": {
      "get": {
        "operationId": "smart_links_document_types_list",
        "description": "Return a list of the selected smart link document types.",
        "parameters": [
          {
            "name": "_ordering",
            "in": "query",
            "description": "Which field to use when ordering the results.",
            "required": false,
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "description": "A page number within the paginated result set.",
            "required": false,
            "type": "integer"
          },
          {
            "name": "page_size",
            "in": "query",
            "description": "Number of results to return per page.",
            "required": false,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "smart_links"
        ]
      },
      "parameters": [
        {
          "name": "smart_link_id",
          "in": "path",
          "required": true,
          "type": "string"
        }
      ]
    },
...
    "/tags/{tag_id}/documents/": {
      "get": {
        "operationId": "tags_documents_list",
        "description": "Returns a list of all the documents tagged by a particular tag.",
        "parameters": [
          {
            "name": "_ordering",
            "in": "query",
            "description": "Which field to use when ordering the results.",
            "required": false,
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "description": "A page number within the paginated result set.",
            "required": false,
            "type": "integer"
          },
          {
            "name": "page_size",
            "in": "query",
            "description": "Number of results to return per page.",
            "required": false,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "tags"
        ]
      },
      "parameters": [
        {
          "name": "tag_id",
          "in": "path",
          "required": true,
          "type": "string"
        }
      ]
    },
...
    "/web_links/{web_link_id}/document_types/": {
      "get": {
        "operationId": "web_links_document_types_list",
        "description": "Return a list of the selected web link document types.",
        "parameters": [
          {
            "name": "_ordering",
            "in": "query",
            "description": "Which field to use when ordering the results.",
            "required": false,
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "description": "A page number within the paginated result set.",
            "required": false,
            "type": "integer"
          },
          {
            "name": "page_size",
            "in": "query",
            "description": "Number of results to return per page.",
            "required": false,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "web_links"
        ]
      },
      "parameters": [
        {
          "name": "web_link_id",
          "in": "path",
          "required": true,
          "type": "string"
        }
      ]
    },
...
    "/workflow_templates/{workflow_template_id}/document_types/": {
      "get": {
        "operationId": "workflow_templates_document_types_list",
        "description": "Returns a list of all the document types attached to a workflow template.",
        "parameters": [
          {
            "name": "_ordering",
            "in": "query",
            "description": "Which field to use when ordering the results.",
            "required": false,
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "description": "A page number within the paginated result set.",
            "required": false,
            "type": "integer"
          },
          {
            "name": "page_size",
            "in": "query",
            "description": "Number of results to return per page.",
            "required": false,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "workflow_templates"
        ]
      },
      "parameters": [
        {
          "name": "workflow_template_id",
          "in": "path",
          "required": true,
          "type": "string"
        }
      ]
    },
...
  },
...
}